# planning/api-documentations/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "echomemory backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/albums/admin": {
      "post": {
        "tags": [
          "albums"
        ],
        "summary": "Create Album",
        "description": "管理员创建专辑。\n\n封面图片通过上传方式提供，服务端自动上传到 OSS 并生成 URL。\n标签由系统根据歌曲收藏自动计算，不允许手动编辑。",
        "operationId": "create_album_api_v1_albums_admin_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_album_api_v1_albums_admin_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/albums/admin/{album_id}": {
      "patch": {
        "tags": [
          "albums"
        ],
        "summary": "Admin Update Album",
        "description": "管理员修改专辑信息（不含文件替换和标签编辑）。",
        "operationId": "admin_update_album_api_v1_albums_admin__album_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlbumUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "albums"
        ],
        "summary": "Admin Delete Album",
        "description": "管理员软删除专辑。",
        "operationId": "admin_delete_album_api_v1_albums_admin__album_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/albums/admin/{album_id}/musics/{music_id}": {
      "post": {
        "tags": [
          "albums"
        ],
        "summary": "Add Music To Album",
        "description": "将一首已上架音乐加入专辑。",
        "operationId": "add_music_to_album_api_v1_albums_admin__album_id__musics__music_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          },
          {
            "name": "music_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Music Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "albums"
        ],
        "summary": "Remove Music From Album",
        "description": "从专辑移除一首音乐。",
        "operationId": "remove_music_from_album_api_v1_albums_admin__album_id__musics__music_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          },
          {
            "name": "music_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Music Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/albums/search": {
      "get": {
        "tags": [
          "albums"
        ],
        "summary": "Search Albums",
        "description": "按标题模糊搜索未删除专辑。",
        "operationId": "search_albums_api_v1_albums_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按标题模糊搜索",
              "title": "Q"
            },
            "description": "按标题模糊搜索"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlbumListOut"
                  },
                  "title": "Response Search Albums Api V1 Albums Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/albums/": {
      "get": {
        "tags": [
          "albums"
        ],
        "summary": "List Albums",
        "description": "分页列出未删除专辑，支持标签筛选。",
        "operationId": "list_albums_api_v1_albums__get",
        "parameters": [
          {
            "name": "emotion_tag_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Emotion Tag Id"
            }
          },
          {
            "name": "interest_tag_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Interest Tag Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlbumListOut"
                  },
                  "title": "Response List Albums Api V1 Albums  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/albums/{album_id}": {
      "get": {
        "tags": [
          "albums"
        ],
        "summary": "Get Album",
        "description": "获取未删除专辑的详情。",
        "operationId": "get_album_api_v1_albums__album_id__get",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/collections/albums/{album_id}": {
      "post": {
        "tags": [
          "collections"
        ],
        "summary": "Collect Album",
        "description": "收藏专辑。已收藏则静默返回。",
        "operationId": "collect_album_api_v1_collections_albums__album_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumCollectionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "collections"
        ],
        "summary": "Uncollect Album",
        "description": "取消收藏专辑。未收藏则静默成功。",
        "operationId": "uncollect_album_api_v1_collections_albums__album_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collections/albums": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "List Album Collections",
        "description": "查询我的收藏专辑列表（按收藏时间倒序）。",
        "operationId": "list_album_collections_api_v1_collections_albums_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlbumCollectionOut"
                  },
                  "title": "Response List Album Collections Api V1 Collections Albums Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlbumAuthorOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "ordinal": {
            "type": "integer",
            "title": "Ordinal"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "nickname",
          "ordinal"
        ],
        "title": "AlbumAuthorOut",
        "description": "专辑作者输出。"
      },
      "AlbumCollectionOut": {
        "properties": {
          "album": {
            "$ref": "#/components/schemas/AlbumListOut"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "album",
          "created_at"
        ],
        "title": "AlbumCollectionOut",
        "description": "专辑收藏项输出。"
      },
      "AlbumListOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "hot": {
            "type": "integer",
            "title": "Hot"
          },
          "play_count": {
            "type": "integer",
            "title": "Play Count"
          },
          "cover_icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Icon Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "hot",
          "play_count",
          "created_at"
        ],
        "title": "AlbumListOut",
        "description": "专辑列表项输出（精简）。"
      },
      "AlbumMusicOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "is_vip": {
            "type": "boolean",
            "title": "Is Vip"
          },
          "hot": {
            "type": "integer",
            "title": "Hot"
          },
          "play_count": {
            "type": "integer",
            "title": "Play Count"
          },
          "cover_icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Icon Url"
          },
          "ordinal": {
            "type": "integer",
            "title": "Ordinal"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "is_vip",
          "hot",
          "play_count",
          "ordinal"
        ],
        "title": "AlbumMusicOut",
        "description": "专辑内嵌套的音乐输出。"
      },
      "AlbumOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "collect_count": {
            "type": "integer",
            "title": "Collect Count"
          },
          "play_count": {
            "type": "integer",
            "title": "Play Count"
          },
          "hot": {
            "type": "integer",
            "title": "Hot"
          },
          "cover_icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Icon Url"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "authors": {
            "items": {
              "$ref": "#/components/schemas/AlbumAuthorOut"
            },
            "type": "array",
            "title": "Authors",
            "default": []
          },
          "musics": {
            "items": {
              "$ref": "#/components/schemas/AlbumMusicOut"
            },
            "type": "array",
            "title": "Musics",
            "default": []
          },
          "emotion_tags": {
            "items": {
              "$ref": "#/components/schemas/TagOut"
            },
            "type": "array",
            "title": "Emotion Tags",
            "default": []
          },
          "interest_tags": {
            "items": {
              "$ref": "#/components/schemas/TagOut"
            },
            "type": "array",
            "title": "Interest Tags",
            "default": []
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "collect_count",
          "play_count",
          "hot",
          "created_at",
          "updated_at"
        ],
        "title": "AlbumOut",
        "description": "专辑详情输出。"
      },
      "AlbumUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "source": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "author_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Ids"
          }
        },
        "type": "object",
        "title": "AlbumUpdate",
        "description": "管理员修改专辑信息的请求体（不含文件和标签编辑）。"
      },
...
      "Body_create_album_api_v1_albums_admin_post": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "source": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "cover_icon": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Cover Icon"
          },
          "cover": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Cover"
          },
          "author_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Author Ids",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "title",
          "cover_icon",
          "cover"
        ],
        "title": "Body_create_album_api_v1_albums_admin_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TagOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "TagOut",
        "description": "标签输出。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}